Installed chakra

+++ backend/controllers/workoutController.js
@@ -1,7 +1,7 @@
 const Workout = require('../models/workoutModel');
 const mongoose = require('mongoose');
 
-// get all workouts
+// Get all workouts
 const getWorkouts = async (req, res) => {
     const workouts = await Workout
         .find({})
@@ -10,17 +10,29 @@ const getWorkouts = async (req, res) => {
     res.status(200).json(workouts);
 }
 
-// get a single workout
+// Get a single workout
 const getWorkout = async (req, res) => {
     const { id } = req.params
 
+    // Check if Id is valid
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ error: "No such workout exists..."})
+    }
+
+    // Find workout
     const workout = await Workout
         .findById(id)
 
+    // Check if workout exists
+    if(!workout) {
+        return res.status(400).json({ error: "No such workout exists..."})
+    }
+
+    // Return status code and workout
     res.status(200).json(workout)
 }
 
-// create a workout
+// Create a workout
 const createWorkout = async (req, res) => {
     const {title, reps, load, sets }  = req.body
 
@@ -38,11 +50,18 @@ const createWorkout = async (req, res) => {
     }
 }
 
+// Delete a workout
 const deleteWorkout = async (req, res) => {
     const { id } = req.params
+
+    // Check if Id is valid
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ error: "No such workout exists..."})
+    }
+
     const workout = await Workout.findOneAndDelete({ _id: id })
 
-    //check if workout exists
+    // Check if workout exists
     if (!workout) {
         return res.status(400).json({ error: ' No such workout exists...'})
     }
@@ -50,12 +69,23 @@ const deleteWorkout = async (req, res) => {
     res.status(200).json(workout)
 }
 
+// Update a workout
 const updateWorkout = async (req, res) => {
     const { id } = req.params
+
+    // Check if Id is valid
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ error: "No such workout exists..."})
+    }
  
     const workout = await Workout.findOneAndUpdate( { _id: id }, {
         ...req.body
     })
+
+    // Check if workout exists
+    if(!workout) {
+        return res.status(400).json({ error: 'No such workout exists...'})
+    }
 
     res.status(200).json(workout)
 }
